Moved from /asserts package
